Move module initialization to registry
